Move in supportedAlgorithms
