+++ components/script/dom/subtlecrypto/rsassa_pkcs1_v1_5_operation.rs
@@ -17,7 +17,7 @@ use crate::dom::bindings::root::DomRoot;
 use crate::dom::cryptokey::{CryptoKey, Handle};
 use crate::dom::globalscope::GlobalScope;
 use crate::dom::subtlecrypto::rsa_common::{self, RsaAlgorithm};
-use crate::dom::subtlecrypto::supported_algorithms::NormalizedAlgorithm;
+use crate::dom::subtlecrypto::NormalizedAlgorithm;
 use crate::dom::subtlecrypto::{
     ALG_SHA1, ALG_SHA256, ALG_SHA384, ALG_SHA512, ExportedKey, KeyAlgorithmAndDerivatives,
     SubtleRsaHashedImportParams, SubtleRsaHashedKeyGenParams,
